fix bug for int in the array

+++ pytorch/src/train_pada.py
@@ -283,7 +283,7 @@ def train(config):
         label_numpy = labels_source.data.cpu().numpy()
         for j in range(int(inputs.size(0) / 2)):
             weight_ad[j] = class_weight[int(label_numpy[j])]
-        weight_ad = weight_ad / torch.max(weight_ad[0:inputs.size(0)/2])
+        weight_ad = weight_ad / torch.max(weight_ad[0:int(inputs.size(0)/2)])
         for j in range(int(inputs.size(0) / 2), inputs.size(0)):
             weight_ad[j] = 1.0            
         transfer_loss = transfer_criterion(features, ad_net, gradient_reverse_layer, \
